+++ dynet/nodes.cc
@@ -453,22 +453,6 @@ void ConstScalarMultiply::backward_dev_impl(const MyDevice & dev,
 DYNET_NODE_INST_DEV_IMPL(ConstScalarMultiply)
 
 template<class MyDevice>
-void Cube::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
-  fx.tvec().device(*dev.edevice) = xs[0]->tvec().cube();
-}
-
-template<class MyDevice>
-void Cube::backward_dev_impl(const MyDevice & dev,
-                             const vector<const Tensor*>& xs,
-                             const Tensor& fx,
-                             const Tensor& dEdf,
-                             unsigned i,
-                             Tensor& dEdxi) const {
-  dEdxi.tvec().device(*dev.edevice) += dEdf.tvec() * xs[0]->tvec().square() * 3.f;
-}
-DYNET_NODE_INST_DEV_IMPL(Cube)
-
-template<class MyDevice>
 void CwiseQuotient::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
   DYNET_ASSERT(xs.size() == 2, "Failed dimension check in CwiseQuotient::forward (cdiv)");
   if(xs[0]->d.bd == xs[1]->d.bd) {
@@ -650,47 +634,6 @@ void ScalarQuotient::backward_dev_impl(const MyDevice & dev,
 }
 DYNET_NODE_INST_DEV_IMPL(ScalarQuotient)
 
-
-template<class MyDevice>
-void DotProduct::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
-  Eigen::array<int, 1> red_axis; red_axis[0] = 0;
-  Eigen::array<int, 2> bcast; bcast[0] = 1; bcast[1] = fx.d.bd;
-  if(fx.d.bd == 1) {
-    fx.t<0>().device(*dev.edevice) = (xs[0]->tvec() * xs[1]->tvec()).sum();
-  } else if(xs[0]->d.bd == xs[1]->d.bd) {
-    fx.tb<0>().device(*dev.edevice) = (xs[0]->tbvec() * xs[1]->tbvec()).sum(red_axis);
-  } else if(xs[0]->d.bd == 1) {
-    fx.tb<0>().device(*dev.edevice) = (xs[0]->tbvec().broadcast(bcast) * xs[1]->tbvec()).sum(red_axis);
-  } else {
-    fx.tb<0>().device(*dev.edevice) = (xs[0]->tbvec() * xs[1]->tbvec().broadcast(bcast)).sum(red_axis);
-  }
-}
-
-template<class MyDevice>
-void DotProduct::backward_dev_impl(const MyDevice & dev,
-                             const vector<const Tensor*>& xs,
-                             const Tensor& fx,
-                             const Tensor& dEdf,
-                             unsigned i,
-                             Tensor& dEdxi) const {
-  if(fx.d.bd == 1) {
-    Eigen::array<int, 1> bcast; bcast[0] = xs[i]->d.batch_size();
-    dEdxi.tvec().device(*dev.edevice) += xs[1-i]->tvec() * dEdf.tvec().broadcast(bcast);
-  } else {
-    Eigen::array<int, 2> bcast; bcast[0] =xs[i]->d.batch_size(); bcast[1] = 1;
-    if(xs[0]->d.bd == xs[1]->d.bd) {
-      dEdxi.tbvec().device(*dev.edevice) += xs[1-i]->tbvec() * dEdf.tbvec().broadcast(bcast);
-    } else if(dEdxi.d.bd == 1) {
-      Eigen::array<int, 1> red_axis; red_axis[0] = 1;
-      dEdxi.tvec().device(*dev.edevice) += (xs[1-i]->tbvec() * dEdf.tbvec().broadcast(bcast)).sum(red_axis);
-    } else {
-      Eigen::array<int, 2> batchcast; batchcast[0] = 1; batchcast[1] = fx.d.bd;
-      dEdxi.tbvec().device(*dev.edevice) += (xs[1-i]->tbvec().broadcast(batchcast) * dEdf.tbvec().broadcast(bcast));
-    }
-  }
-}
-DYNET_NODE_INST_DEV_IMPL(DotProduct)
-
 template<class MyDevice>
 void Dropout::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
   Tensor m(dim, (float*)aux_mem, fx.device, DeviceMempool::FXS);
@@ -777,22 +720,6 @@ void Erf::backward_dev_impl(const MyDevice & dev,
 DYNET_NODE_INST_DEV_IMPL(Erf)
 
 template<class MyDevice>
-void Exp::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
-  fx.tvec().device(*dev.edevice) = xs[0]->tvec().exp();
-}
-
-template<class MyDevice>
-void Exp::backward_dev_impl(const MyDevice & dev,
-                             const vector<const Tensor*>& xs,
-                             const Tensor& fx,
-                             const Tensor& dEdf,
-                             unsigned i,
-                             Tensor& dEdxi) const {
-  dEdxi.tvec().device(*dev.edevice) += dEdf.tvec() * fx.tvec();
-}
-DYNET_NODE_INST_DEV_IMPL(Exp)
-
-template<class MyDevice>
 void GaussianNoise::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
   Tensor m(dim, (float*)aux_mem, fx.device, DeviceMempool::FXS);
   TensorTools::randomize_normal(m, 0, stddev);
@@ -809,24 +736,6 @@ void GaussianNoise::backward_dev_impl(const MyDevice & dev,
   dEdxi.tvec().device(*dev.edevice) += dEdf.tvec();
 }
 DYNET_NODE_INST_DEV_IMPL(GaussianNoise)
-
-template<class MyDevice>
-void HuberDistance::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
-  DYNET_ASSERT(xs.size() == 2, "HuberDistance::forward dimension check failed");
-  fx.t<0>().device(*dev.edevice) = (xs[0]->tvec() - xs[1]->tvec()).unaryExpr(FHuberForward(d)).sum();
-}
-
-template<class MyDevice>
-void HuberDistance::backward_dev_impl(const MyDevice & dev,
-                             const vector<const Tensor*>& xs,
-                             const Tensor& fx,
-                             const Tensor& dEdf,
-                             unsigned i,
-                             Tensor& dEdxi) const {
-  DYNET_ASSERT(i < 2, "HuberDistance::backward dimension check failed");
-  dEdxi.tvec().device(*dev.edevice) += (xs[i]->tvec() - xs[1-i]->tvec()).unaryExpr(FHuberBackward(d, as_scalar(dEdf)));
-}
-DYNET_NODE_INST_DEV_IMPL(HuberDistance)
 
 template<class MyDevice>
 void Identity::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
@@ -879,40 +788,6 @@ void KMHNGram::backward_dev_impl(const MyDevice & dev,
 #endif
 }
 DYNET_NODE_INST_DEV_IMPL(KMHNGram)
-
-template<class MyDevice>
-void L1Distance::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
-  DYNET_ASSERT(xs.size() == 2, "Failed dimension check in L1Distance::forward");
-  fx.t<0>().device(*dev.edevice) = (xs[0]->tvec() - xs[1]->tvec()).abs().sum();
-}
-
-template<class MyDevice>
-void L1Distance::backward_dev_impl(const MyDevice & dev,
-                             const vector<const Tensor*>& xs,
-                             const Tensor& fx,
-                             const Tensor& dEdf,
-                             unsigned i,
-                             Tensor& dEdxi) const {
-  DYNET_ASSERT(i < 2, "Failed dimension check in L1Distance::backward");
-  dEdxi.tvec().device(*dev.edevice) += (xs[i]->tvec() - xs[1-i]->tvec()).unaryExpr(FL1Backward(as_scalar(dEdf)));
-}
-DYNET_NODE_INST_DEV_IMPL(L1Distance)
-
-template<class MyDevice>
-void Log::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
-  fx.tvec().device(*dev.edevice) = xs[0]->tvec().log();
-}
-
-template<class MyDevice>
-void Log::backward_dev_impl(const MyDevice & dev,
-                             const vector<const Tensor*>& xs,
-                             const Tensor& fx,
-                             const Tensor& dEdf,
-                             unsigned i,
-                             Tensor& dEdxi) const {
-  dEdxi.tvec().device(*dev.edevice) += dEdf.tvec() / xs[0]->tvec();
-}
-DYNET_NODE_INST_DEV_IMPL(Log)
 
 template<class MyDevice>
 void LogDet::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
@@ -1235,24 +1110,6 @@ void Min::backward_dev_impl(const MyDevice & dev,
   }
 }
 DYNET_NODE_INST_DEV_IMPL(Min)
-
-template<class MyDevice>
-void Negate::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
-  DYNET_ASSERT(xs.size() == 1, "Failed dimension check in Negate::forward");
-  fx.tvec().device(*dev.edevice) = -xs[0]->tvec();
-}
-
-template<class MyDevice>
-void Negate::backward_dev_impl(const MyDevice & dev,
-                             const vector<const Tensor*>& xs,
-                             const Tensor& fx,
-                             const Tensor& dEdf,
-                             unsigned i,
-                             Tensor& dEdxi) const {
-  DYNET_ASSERT(i == 0, "Failed dimension check in Negate::backward");
-  dEdxi.tvec().device(*dev.edevice) -= dEdf.tvec();
-}
-DYNET_NODE_INST_DEV_IMPL(Negate)
 
 template<class MyDevice>
 void PairwiseRankLoss::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
@@ -1745,145 +1602,6 @@ void SparsemaxLoss::backward_dev_impl(const MyDevice & dev,
 #endif
 }
 DYNET_NODE_INST_DEV_IMPL(SparsemaxLoss)
-
-template<class MyDevice>
-void Square::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
-  fx.tvec().device(*dev.edevice) = xs[0]->tvec().square();
-}
-
-template<class MyDevice>
-void Square::backward_dev_impl(const MyDevice & dev,
-                             const vector<const Tensor*>& xs,
-                             const Tensor& fx,
-                             const Tensor& dEdf,
-                             unsigned i,
-                             Tensor& dEdxi) const {
-  dEdxi.tvec().device(*dev.edevice) += dEdf.tvec() * xs[0]->tvec() * 2.f;
-}
-DYNET_NODE_INST_DEV_IMPL(Square)
-
-template<class MyDevice>
-void SquaredEuclideanDistance::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
-  DYNET_ASSERT(xs.size() == 2, "Failed dimension check in SquaredEuclideanDistance::forward");
-  Eigen::array<ptrdiff_t, 1> red_axis = {0};
-  if(xs[0]->d.bd == xs[1]->d.bd) {
-    fx.tb<0>().device(*dev.edevice) = (xs[0]->tbvec() - xs[1]->tbvec()).square().sum(red_axis);
-  } else if(xs[0]->d.bd == 1) {
-    Eigen::array<ptrdiff_t, 2> bcast = {1, xs[1]->d.bd};
-    fx.tb<0>().device(*dev.edevice) = (xs[0]->tbvec().broadcast(bcast) - xs[1]->tbvec()).square().sum(red_axis);
-  } else {
-    Eigen::array<ptrdiff_t, 2> bcast = {1, xs[0]->d.bd};
-    fx.tb<0>().device(*dev.edevice) = (xs[0]->tbvec() - xs[1]->tbvec().broadcast(bcast)).square().sum(red_axis);
-  }
-}
-
-template<class MyDevice>
-void SquaredEuclideanDistance::backward_dev_impl(const MyDevice & dev,
-                             const vector<const Tensor*>& xs,
-                             const Tensor& fx,
-                             const Tensor& dEdf,
-                             unsigned i,
-                             Tensor& dEdxi) const {
-  DYNET_ASSERT(i < 2, "Failed dimension check in SquaredEuclideanDistance::backward");
-  float multiplier = (i == 1 ? -2.0f : 2.0f);
-  Eigen::array<ptrdiff_t, 2> bcast = {xs[0]->d.batch_size(), 1};
-  if(xs[0]->d.bd == xs[1]->d.bd) {
-    dEdxi.tbvec().device(*dev.edevice) += (xs[0]->tbvec() - xs[1]->tbvec()) * dEdf.tbvec().broadcast(bcast) * multiplier;
-  } else if(xs[0]->d.bd == 1) {
-    Eigen::array<ptrdiff_t, 2> batchcast = {1, xs[1]->d.bd};
-    if(i == 1) {
-      dEdxi.tbvec().device(*dev.edevice) += (xs[0]->tbvec().broadcast(batchcast) - xs[1]->tbvec()) * dEdf.tbvec().broadcast(bcast) * multiplier;
-    } else {
-      Eigen::array<ptrdiff_t, 1> red_axis = {1};
-      dEdxi.tvec().device(*dev.edevice) += ((xs[0]->tbvec().broadcast(batchcast) - xs[1]->tbvec()) * dEdf.tbvec().broadcast(bcast) * multiplier).sum(red_axis);
-    }
-  } else {
-    Eigen::array<ptrdiff_t, 2> batchcast = {1, xs[0]->d.bd};
-    if(i == 0) {
-      dEdxi.tbvec().device(*dev.edevice) += (xs[0]->tbvec() - xs[1]->tbvec().broadcast(batchcast)) * dEdf.tbvec().broadcast(bcast) * multiplier;
-    } else {
-      Eigen::array<ptrdiff_t, 1> red_axis = {1};
-      dEdxi.tvec().device(*dev.edevice) += ((xs[0]->tbvec() - xs[1]->tbvec().broadcast(batchcast)) * dEdf.tbvec().broadcast(bcast) * multiplier).sum(red_axis);
-    }
-  }
-}
-DYNET_NODE_INST_DEV_IMPL(SquaredEuclideanDistance)
-
-template<class MyDevice>
-void SquaredNorm::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
-  DYNET_ASSERT(xs.size() == 1, "Failed dimension check in SquaredNorm::forward");
-  Eigen::array<ptrdiff_t, 1> red_axis = {0};
-  fx.tb<0>().device(*dev.edevice) = xs[0]->tbvec().square().sum(red_axis);
-}
-
-template<class MyDevice>
-void SquaredNorm::backward_dev_impl(const MyDevice & dev,
-                             const vector<const Tensor*>& xs,
-                             const Tensor& fx,
-                             const Tensor& dEdf,
-                             unsigned i,
-                             Tensor& dEdxi) const {
-  DYNET_ASSERT(i < 1, "Failed dimension check in SquaredNorm::backward");
-  Eigen::array<ptrdiff_t, 2> bcast = {xs[0]->d.batch_size(), 1};
-  dEdxi.tbvec().device(*dev.edevice) += xs[0]->tbvec() * dEdf.tbvec().broadcast(bcast) * 2.0f;
-}
-DYNET_NODE_INST_DEV_IMPL(SquaredNorm)
-
-template<class MyDevice>
-void L2Norm::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
-  DYNET_ASSERT(xs.size() == 1, "Failed dimension check in L2Norm::forward");
-  Eigen::array<ptrdiff_t, 1> red_axis = {0};
-  fx.tb<0>().device(*dev.edevice) =
-      (xs[0]->tbvec().square().sum(red_axis)).sqrt();
-}
-
-template<class MyDevice>
-void L2Norm::backward_dev_impl(const MyDevice & dev,
-                             const vector<const Tensor*>& xs,
-                             const Tensor& fx,
-                             const Tensor& dEdf,
-                             unsigned i,
-                             Tensor& dEdxi) const {
-  DYNET_ASSERT(i < 1, "Failed dimension check in L2Norm::backward");
-  Eigen::array<ptrdiff_t, 2> bcast = {xs[0]->d.batch_size(), 1};
-  dEdxi.tbvec().device(*dev.edevice) +=
-      xs[0]->tbvec() *
-      (2 * fx.tbvec().binaryExpr(dEdf.tbvec(),
-                                 FSqrtBackward())).broadcast(bcast);
-}
-DYNET_NODE_INST_DEV_IMPL(L2Norm)
-
-template<class MyDevice>
-void Sqrt::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
-  fx.tvec().device(*dev.edevice) = xs[0]->tvec().sqrt();
-}
-
-template<class MyDevice>
-void Sqrt::backward_dev_impl(const MyDevice & dev,
-                             const vector<const Tensor*>& xs,
-                             const Tensor& fx,
-                             const Tensor& dEdf,
-                             unsigned i,
-                             Tensor& dEdxi) const {
-  dEdxi.tvec().device(*dev.edevice) += fx.tvec().binaryExpr(dEdf.tvec(), FSqrtBackward());
-}
-DYNET_NODE_INST_DEV_IMPL(Sqrt)
-
-template<class MyDevice>
-void Abs::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
-  fx.tvec().device(*dev.edevice) = xs[0]->tvec().abs();
-}
-
-template<class MyDevice>
-void Abs::backward_dev_impl(const MyDevice & dev,
-                             const vector<const Tensor*>& xs,
-                             const Tensor& fx,
-                             const Tensor& dEdf,
-                             unsigned i,
-                             Tensor& dEdxi) const {
-  dEdxi.tvec().device(*dev.edevice) += dEdf.tvec() * xs[0]->tvec().sign();
-}
-DYNET_NODE_INST_DEV_IMPL(Abs)
 
 template<class MyDevice>
 void Sum::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
